pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "tunadex"
version = "2.0.0"
description = "Automated seafood shipment tracking from email"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "gspread>=6.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "pydantic>=2.5.0",
    "pdfplumber>=0.10.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "Jinja2>=3.1.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
tunedex = "tunadex.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
